Baseline for DoT integration into sdig, taken from Habbie/sdig-dot-pin
