+++ pdns/dnsdistdist/m4/pdns_with_gnutls.m4
@@ -1,0 +1,1 @@
+../../../m4/pdns_with_gnutls.m4
